c) testing

+++ 2/Blatt2/src/b_c/Match_Object.java
@@ -8,11 +8,11 @@ public class Match_Object {
     
     private final String protein_ID;
     private final String source_DB;
-    private final float e_value;
+    private final double e_value;
     private final int score;
     private final int round;
     
-    public Match_Object(String protein_ID, String source_DB, float e_value, int score, int round){
+    public Match_Object(String protein_ID, String source_DB, double e_value, int score, int round){
         this.protein_ID = protein_ID;
         this.source_DB = source_DB;
         this.e_value = e_value;
